+++ src/algorithms/mod.rs
@@ -55,6 +55,7 @@ pub use utils::IdentifyDistinct;
 #[doc(no_inline)]
 pub use crate::Algorithm;
 
+pub mod histogram;
 pub mod hunt;
 pub mod lcs;
 pub mod myers;
@@ -115,6 +116,9 @@ where
         Algorithm::Patience => patience::diff_deadline(d, old, old_range, new, new_range, deadline),
         Algorithm::Lcs => lcs::diff_deadline(d, old, old_range, new, new_range, deadline),
         Algorithm::Hunt => hunt::diff_deadline(d, old, old_range, new, new_range, deadline),
+        Algorithm::Histogram => {
+            histogram::diff_deadline(d, old, old_range, new, new_range, deadline)
+        }
     }
 }